Moved random methods from FlightPlanGnerator to Utils
