+++ src/main/java/seprhou/gui/GameOverScreen.java
@@ -3,6 +3,7 @@ package seprhou.gui;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.scenes.scene2d.ui.Image;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
+import seprhou.logic.Utils;
 
 /**
  * The screen displayed after the game is over
