+++ src/main/java/seprhou/logic/FlightPlanGenerator.java
@@ -54,58 +54,6 @@ public class FlightPlanGenerator
 	public void setMaxWaypoints(int maxWaypoints) { this.maxWaypoints = maxWaypoints; }
 
 	/**
-	 * Generates a random subset of the given list
-	 *
-	 * @param list list of items
-	 * @param n number of items to choose
-	 * @param <T> type of the items in the list
-	 * @return the chosen item
-	 */
-	private <T> List<T> randomSubset(List<T> list, int n)
-	{
-		// Copy the list and shuffle it
-		List<T> result = new ArrayList<>(list);
-		Collections.shuffle(result, random);
-
-		// Return first n items
-		return result.subList(0, n);
-	}
-
-	/**
-	 * Chooses an item from a list
-	 *
-	 * @param list list to choose from
-	 * @param <T> type of the items in the list
-	 * @return the chosen item
-	 */
-	private <T> T randomItem(List<T> list)
-	{
-		return list.get(random.nextInt(list.size()));
-	}
-
-	/**
-	 * Chooses an item from a list but does not include the invalidItem
-	 *
-	 * @param list list to choose from
-	 * @param invalidItem the item which will not be picked
-	 * @param <T> type of the items in the list
-	 * @return the chosen item
-	 */
-	private <T> T randomItem(List<T> list, T invalidItem)
-	{
-		// Check for impossible situation
-		if (list.size() == 1 && list.get(0) == invalidItem)
-			throw new IllegalStateException("list given to randomItem contains no valid items!");
-
-		// Choose one item and skip it if it's invalid
-		int item = random.nextInt(list.size());
-		if (list.get(item) == invalidItem)
-			item = (item + 1) % list.size();
-
-		return list.get(item);
-	}
-
-	/**
 	 * Generates the list of points which a flight can safely enter currently
 	 *
 	 * @param airspace airspace to gather data from
@@ -164,17 +112,17 @@ public class FlightPlanGenerator
 
 		// Choose some waypoints + 2 entry and exit points
 		int waypointCount = random.nextInt(maxWaypoints - minWaypoints) + minWaypoints;
-		List<Vector2D> myWaypoints = randomSubset(waypoints, waypointCount);
-		Vector2D entryPoint = randomItem(entryPointSubset);
-		Vector2D exitPoint = randomItem(entryExitPoints, entryPoint);
+		List<Vector2D> myWaypoints = Utils.randomSubset(waypoints, waypointCount);
+		Vector2D entryPoint = Utils.randomItem(entryPointSubset);
+		Vector2D exitPoint = Utils.randomItem(entryExitPoints, entryPoint);
 
 		// Insert entry + exit points into the list
 		myWaypoints.add(0, entryPoint);
 		myWaypoints.add(exitPoint);
 
 		// Choose initial speed and altitude
-		float initialSpeed = randomItem(initialSpeeds);
-		float initialAltitude = randomItem(initialAltitudes);
+		float initialSpeed = Utils.randomItem(initialSpeeds);
+		float initialAltitude = Utils.randomItem(initialAltitudes);
 
 		// Create flight plan
 		return new FlightPlan(myWaypoints, initialSpeed, initialAltitude);
